refactor(quotes): move common types and errors into common package
