+++ pkg/quotes/interface.go
@@ -2,12 +2,9 @@
 package quotes
 
 import (
-	"errors"
 	"fmt"
-	"time"
 
 	"github.com/ipfs/go-log/v2"
-	"github.com/shopspring/decimal"
 )
 
 var logger = log.Logger("quotes")
@@ -34,39 +31,3 @@ func NewDriver(config Config, outbox chan<- TradeEvent) (Driver, error) {
 	}
 	return driver, nil
 }
-
-type Config struct {
-	URL             string             `yaml:"url" env:"QUOTES_URL" env-default:""`
-	Driver          DriverType         `yaml:"driver" env:"QUOTES_DRIVER" env-default:"binance"`
-	ReconnectPeriod time.Duration      `yaml:"period" env:"QUOTES_RECONNECT_PERIOD" env-default:"5s"`
-	TradeSampler    TradeSamplerConfig `yaml:"trade_sampler"`
-}
-
-type TradeSamplerConfig struct {
-	Enabled           bool `yaml:"enabled" env:"QUOTES_TRADE_SAMPLER_ENABLED"`
-	DefaultPercentage int  `yaml:"default_percentage" env:"QUOTES_TRADE_SAMPLER_DEFAULT_PERCENTAGE"`
-}
-
-type Market struct {
-	BaseUnit  string // e.g. `btc` in `btcusdt`
-	QuoteUnit string // e.g. `usdt` in `btcusdt`
-}
-
-// TradeEvent is a generic container
-// for trades received from providers.
-type TradeEvent struct {
-	Source    DriverType
-	Market    string // e.g. `btcusdt`
-	Price     decimal.Decimal
-	Amount    decimal.Decimal
-	Total     decimal.Decimal
-	TakerType TakerType
-	CreatedAt time.Time
-}
-
-var (
-	ErrNotSubbed     = errors.New("market not subscribed")
-	ErrAlreadySubbed = errors.New("market already subscribed")
-	ErrFailedSub     = errors.New("failed to subscribe to market")
-	ErrFailedUnsub   = errors.New("failed to unsubscribe from market")
-)
